added 1.20.5 and 1.20.6 vanilla support

+++ api/src/main/java/de/funboyy/addon/worldedit/cui/api/MinecraftHelper.java
@@ -9,7 +9,7 @@ public interface MinecraftHelper {
 
   HitResult pick(final Entity entity, final double distance, final float partialTicks, final boolean fluids);
 
-  float getPickDistance();
+  double getPickDistance();
 
   void pushProfiler(final String name);
