+++ game-runner/src/v1_19_4/java/de/funboyy/addon/worldedit/cui/v1_19_4/VersionedMinecraftHelper.java
@@ -17,7 +17,7 @@ public class VersionedMinecraftHelper implements MinecraftHelper {
   }
 
   @Override
-  public float getPickDistance() {
+  public double getPickDistance() {
     return this.minecraft.gameMode.getPickRange();
   }
